rename new test file
