+++ packages/syft/tests/syft/service/sync/sync_flow_test.py
@@ -248,43 +248,6 @@ def test_sync_flow():
     high_worker.cleanup()
 
 
-def test_diff_state(low_worker, high_worker):
-    low_client: DomainClient = low_worker.root_client
-    client_low_ds = get_ds_client(low_client)
-    high_client: DomainClient = high_worker.root_client
-
-    @sy.syft_function_single_use()
-    def compute() -> int:
-        return 42
-
-    compute.code = dedent(compute.code)
-
-    _ = client_low_ds.code.request_code_execution(compute)
-
-    diff_state_before, diff_state_after = compare_and_resolve(
-        from_client=low_client, to_client=high_client
-    )
-
-    assert not diff_state_before.is_same
-
-    assert diff_state_after.is_same
-
-    run_and_accept_result(high_client)
-    diff_state_before, diff_state_after = compare_and_resolve(
-        from_client=high_client, to_client=low_client
-    )
-
-    high_state = high_client.get_sync_state()
-    low_state = high_client.get_sync_state()
-    assert high_state.get_previous_state_diff().is_same
-    assert low_state.get_previous_state_diff().is_same
-    assert diff_state_after.is_same
-
-    client_low_ds.refresh()
-    res = client_low_ds.code.compute(blocking=True)
-    assert res == compute(blocking=True).get()
-
-
 def test_forget_usercode(low_worker, high_worker):
     low_client = low_worker.root_client
     client_low_ds = low_worker.guest_client
@@ -330,69 +293,6 @@ def test_forget_usercode(low_worker, high_worker):
         share_private_objects=True,
         decision_callback=skip_if_user_code,
     )
-
-
-@sy.mock_api_endpoint()
-def mock_function(context) -> str:
-    return -42
-
-
-@sy.private_api_endpoint()
-def private_function(context) -> str:
-    return 42
-
-
-def test_twin_api_integration(low_worker, high_worker):
-    low_client = low_worker.root_client
-    high_client = high_worker.root_client
-    client_low_ds = get_ds_client(low_client)
-
-    new_endpoint = sy.TwinAPIEndpoint(
-        path="testapi.query",
-        private_function=private_function,
-        mock_function=mock_function,
-        description="",
-    )
-    high_client.api.services.api.add(endpoint=new_endpoint)
-    high_client.refresh()
-    high_private_res = high_client.api.services.testapi.query.private()
-    assert high_private_res == 42
-
-    diff_before, diff_after = compare_and_resolve(
-        from_client=high_client, to_client=low_client
-    )
-
-    client_low_ds.refresh()
-    low_private_res = client_low_ds.api.services.testapi.query.private()
-    assert isinstance(
-        low_private_res, SyftError
-    ), "Should not have access to private on low side"
-    low_mock_res = client_low_ds.api.services.testapi.query.mock()
-    high_mock_res = high_client.api.services.testapi.query.mock()
-    assert low_mock_res == high_mock_res == -42
-
-    @syft.syft_function_single_use(
-        query=client_low_ds.api.services.testapi.query,
-    )
-    def compute(query):
-        return query()
-
-    compute.code = dedent(compute.code)
-    _ = client_low_ds.code.request_code_execution(compute)
-
-    diff_before, diff_after = compare_and_resolve(
-        from_client=low_client, to_client=high_client
-    )
-
-    job_high = high_client.code.compute(query=high_client.api.services.testapi.query)
-    high_client.requests[0].accept_by_depositing_result(job_high)
-    diff_before, diff_after = compare_and_resolve(
-        from_client=high_client, to_client=low_client
-    )
-    client_low_ds.refresh()
-    res = client_low_ds.code.compute(query=client_low_ds.api.services.testapi.query)
-    assert res.syft_action_data == high_client.api.services.testapi.query.private()
-    assert diff_after.is_same
 
 
 def test_skip_user_code(low_worker, high_worker):
